LayoutTests/navigation-api/navigation-object-is-replaceable.html

<script src="../resources/js-test.js"></script>

<script>
description("This tests that the Navigation object is replaceable");

window.onload = () => {
    shouldNotBe("navigation", "\"foo\"");
    navigation = "foo";
    shouldBe("navigation", "\"foo\"");
}

</script>
