LayoutTests/navigation-api/navigation-object-is-replaceable-expected.txt

This tests that the Navigation object is replaceable

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigation is not "foo"
PASS navigation is "foo"
PASS successfullyParsed is true

TEST COMPLETE
